backend/ai-service/requirements.txt
openai==1.3.8
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
redis[hiredis]==5.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.2
